requirements.txt
flask
requests
slack_sdk
transformers
torch
python-dotenv
openai
sentence-transformers